requirements.txt
altair==6.0.0
attrs==25.4.0
audioread==3.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
msgpack==1.1.2
narwhals==2.15.0
numba==0.63.1
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
praat-parselmouth==0.4.7
protobuf==6.33.5
pyarrow==23.0.0
pycparser==3.0
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==1.0.0
streamlit==1.53.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0